requirements.txt
attrs==23.1.0
joblib==1.2.0
numpy==1.24.3
pyrsistent==0.19.3
torch==2.0.1
certifi==2023.5.7
streamlit-faker==0.0.2
urllib3==2.0.3
nvidia-cufft-cu11==10.9.0.58
click==8.1.3
scikit-learn==1.2.2
packaging==23.1
streamlit-camera-input-live==0.2.0
torchutils==0.0.4
pandas==2.0.2
tqdm==4.65.0
charset-normalizer==3.1.0
torchmetrics==0.11.4
gitdb==4.0.10
nvidia-cudnn-cu11==8.5.0.96
tornado==6.3.2
zipp==3.15.0
decorator==5.1.1
tokenizers==0.13.3
torchaudio==2.0.2
idna==3.4
cachetools==5.3.1
MarkupSafe==2.1.3
pytz==2023.3
fonttools==4.39.4
toml==0.10.2
lit==16.0.5.post0
gensim==4.3.1
python-dateutil==2.8.2
smart-open==6.3.0
tzdata==2023.3
jsonschema==4.17.3
nvidia-nccl-cu11==2.14.3
pytz-deprecation-shim==0.1.0.post0
Pillow==9.5.0
triton==2.0.0
streamlit-toggle-switch==1.0.2
networkx==3.1
mpmath==1.3.0
fsspec==2023.5.0
streamlit==1.23.1
pydeck==0.8.1b0
huggingface-hub==0.15.1
streamlit-image-coordinates==0.1.3
toolz==0.12.0
protobuf==4.23.2
streamlit-keyup==0.2.0
matplotlib==3.7.1
requests==2.31.0
filelock==3.12.0
blinker==1.6.2
Pympler==1.0.1
Jinja2==3.1.2
streamlit-embedcode==0.1.2
pyarrow==12.0.0
markdown-it-py==2.2.0
contourpy==1.0.7
smmap==5.0.0
streamlit-card==0.0.5
typing_extensions==4.6.3
nvidia-cuda-nvrtc-cu11==11.7.99
altair==5.0.1
cycler==0.11.0
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cusolver-cu11==11.4.0.1
mdurl==0.1.2
tenacity==8.2.2
nvidia-curand-cu11==10.2.10.91
cmake==3.26.4
nltk==3.8.1
six==1.16.0
tzlocal==4.3
streamlit-extras==0.2.7
nvidia-nvtx-cu11==11.7.91
nvidia-cublas-cu11==11.10.3.66
scipy==1.10.1
PyYAML==6.0
validators==0.20.0
importlib-metadata==6.6.0
threadpoolctl==3.1.0
sympy==1.12
watchdog==3.0.0
streamlit-vertical-slider==1.0.2
pyparsing==3.0.9
kiwisolver==1.4.4
rich==13.4.1
torchvision==0.15.2
nvidia-cusparse-cu11==11.7.4.91
Pygments==2.15.1
nvidia-cuda-cupti-cu11==11.7.101
GitPython==3.1.31
transformers==4.29.2
plotly==5.14.1
regex==2023.6.3
